:sparkles: feat: 기기 재부팅 시 알람 재등록 서비스에서 처리
Issue #86
